refactor: update monitor dashboard visuals with new background images and adjust styles for improved layout and aesthetics
